+++ src/scheduler.cc
@@ -35,7 +35,7 @@ void Scheduler::sjf (vector<Process *> &q) {
         // seta a duracao como prioridade
         new_p -> set_priority(new_p -> get_duration());
         // percorre os processos do vetor processes
-        for (int i = 0; i < q.size(); i++) {
+        for (long unsigned int i = 0; i < q.size(); i++) {
             // verifica se a prioridade do prox processo e menor
             if (new_p -> get_priority() < q[i] -> get_priority()) {
                 // retira o novo processo do vetor "fila"
@@ -52,7 +52,7 @@ void Scheduler::sjf (vector<Process *> &q) {
         // declara variavel auxiliar "priority"
         int priority;
         // percorre os processos do vetor processes
-        for (int i = 0; i < q.size(); i++) {
+        for (long unsigned int i = 0; i < q.size(); i++) {
             // verifica se o processo nao e o escolhido no escalonamento
             if (q[i] -> get_pid() != q.front() -> get_pid()) {
                 // aumenta a prioeidade do processo
@@ -78,7 +78,7 @@ void Scheduler::preemptive_prio(vector<Process *> &q) {
         // seta o estado do processo como "ready", "pronto"
         new_p -> set_state_ready();
         // percorre os processos do vetor "fila" de processes
-        for (int i = 0; i < q.size(); i++) {
+        for (long unsigned int i = 0; i < q.size(); i++) {
             // verifica se a prioridade do novo processo e maior que a do processo de posicao i
             if (new_p -> get_priority() > q[i] -> get_priority()) {
                 // retira o novo processo do vetor "fila"
@@ -97,7 +97,7 @@ void Scheduler::preemptive_prio(vector<Process *> &q) {
         // declara variavel auxiliar "priority"
         int priority;
         // percorre os processos do vetor processes
-        for (int i = 1; i < q.size(); i++) {
+        for (long unsigned int i = 1; i < q.size(); i++) {
             // aumenta a prioeidade do processo
             priority = q[i] -> get_priority() - 1;
             q[i] -> set_priority(priority);
@@ -119,7 +119,7 @@ void Scheduler::non_preemptive_prio(vector<Process *> &q) {
         // seta o estado do processo como "ready", "pronto"
         new_p -> set_state_ready();
         // percorre os processos do vetor "fila" de processes
-        for (int i = 1; i < q.size(); i++) {
+        for (long unsigned int i = 1; i < q.size(); i++) {
             // verifica se a prioridade do novo processo e maior que a do processo de posicao i
             if (new_p -> get_priority() > q[i] -> get_priority()) {
                 // retira o novo processo do vetor "fila"
@@ -132,7 +132,7 @@ void Scheduler::non_preemptive_prio(vector<Process *> &q) {
         // declara variavel auxiliar "priority"
         int priority;
         // percorre os processos do vetor processes
-        for (int i = 1; i < q.size(); i++) {
+        for (long unsigned int i = 1; i < q.size(); i++) {
             // aumenta a prioeidade do processo
             priority = q[i] -> get_priority() - 1;
             q[i] -> set_priority(priority);
